+++ packages/connectors/hull-zapier/test/unit/receive-subscription.spec.js
@@ -19,21 +19,319 @@ describe("Zapier Subscription Tests", () => {
     "");
 
 
-  it("Receive New Subscription", () => {
+  it("Receive User Entered Segment Subscription", () => {
     const subscribe = _.cloneDeep(testDefinition);
     subscribe.input.data = {
       body: {
         "url": "https://hooks.zapier.com/hooks/standard/1/1/",
         "action": "entered_segment",
-        "entityType": "user"
+        "entityType": "user",
+        "inputData": {
+          "user_segments": ["all_segments"]
+        }
       }
     };
     subscribe.serviceRequests[0].input = {
-      "subscriptions": [
+      "triggers": [
         {
-          "url": "https://hooks.zapier.com/hooks/standard/1/1/",
-          "action": "entered_segment",
-          "entityType": "user"
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "entered_user_segments": ["all_segments"]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive User Left Segment Subscription", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "left_segment",
+        "entityType": "user",
+        "inputData": {
+          "user_segments": ["all_segments"]
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "left_user_segments": ["all_segments"]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive User Attribute Updated Subscription Subscription With All Valid Fields", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "attribute_updated",
+        "entityType": "user",
+        "inputData": {
+          "user_segments": [ "segment_1" ],
+          "account_segments": [ "account_segment_1" ],
+          "user_attributes": [ "pipedrive/department" ],
+          "account_attributes": [ "pipedrive/industry", "num_employees" ]
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "user_segments": [ "segment_1" ],
+            "account_segments": [ "account_segment_1" ],
+            "user_attribute_updated": [ "pipedrive/department" ],
+            "account_attribute_updated": [ "pipedrive/industry", "num_employees" ]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive User Attribute Updated Subscription Subscription With No Account Attributes", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "attribute_updated",
+        "entityType": "user",
+        "inputData": {
+          "user_segments": [ "segment_1" ],
+          "account_segments": [ "account_segment_1" ],
+          "user_attributes": [ "pipedrive/department" ],
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "user_segments": [ "segment_1" ],
+            "account_segments": [ "account_segment_1" ],
+            "user_attribute_updated": [ "pipedrive/department" ]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive User Attribute Updated Subscription Subscription With No User Attributes", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "attribute_updated",
+        "entityType": "user",
+        "inputData": {
+          "user_segments": [ "segment_1" ],
+          "account_segments": [ "account_segment_1" ],
+          "account_attributes": [ "pipedrive/industry", "num_employees" ]
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "user_segments": [ "segment_1" ],
+            "account_segments": [ "account_segment_1" ],
+            "account_attribute_updated": [ "pipedrive/industry", "num_employees" ]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive User Event Created Subscription Subscription", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "created",
+        "entityType": "user_event",
+        "inputData": {
+          "user_segments": [ "segment_1" ],
+          "account_segments": [ "account_segment_1" ],
+          "user_events": [ "Email Opened" ]
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "user_segments": [ "segment_1" ],
+            "account_segments": [ "account_segment_1" ],
+            "user_events": [ "Email Opened" ]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive User Created Subscription Subscription", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "created",
+        "entityType": "user",
+        "inputData": {
+          "user_segments": [ "segment_1" ]
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "is_new_user": true,
+            "user_segments": [ "segment_1" ]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive Account Entered Segment Subscription", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "entered_segment",
+        "entityType": "account",
+        "inputData": {
+          "account_segments": ["all_segments"]
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "entered_account_segments": ["all_segments"]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive Account Left Segment Subscription", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "left_segment",
+        "entityType": "account",
+        "inputData": {
+          "account_segments": ["all_segments"]
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "left_account_segments": ["all_segments"]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive Account Attribute Updated Subscription Subscription", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "attribute_updated",
+        "entityType": "account",
+        "inputData": {
+          "account_segments": [ "account_segment_1" ],
+          "account_attributes": [ "pipedrive/industry", "num_employees" ]
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "account_segments": [ "account_segment_1" ],
+            "account_attribute_updated": [ "pipedrive/industry", "num_employees" ]
+          }
+        }
+      ]
+    };
+    return harness.runTest(subscribe);
+  });
+
+  it("Receive Account Created Subscription Subscription", () => {
+    const subscribe = _.cloneDeep(testDefinition);
+    subscribe.input.data = {
+      body: {
+        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+        "action": "created",
+        "entityType": "account",
+        "inputData": {
+          "account_segments": [ "segment_1" ]
+        }
+      }
+    };
+    subscribe.serviceRequests[0].input = {
+      "triggers": [
+        {
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "is_new_account": true,
+            "account_segments": [ "segment_1" ]
+          }
         }
       ]
     };
@@ -42,31 +340,43 @@ describe("Zapier Subscription Tests", () => {
 
   it("Receive New Subscription And Merge With Existing Subscriptions", () => {
     const subscribe = _.cloneDeep(testDefinition);
-    subscribe.configuration.private_settings.subscriptions = [
+    subscribe.configuration.private_settings.triggers = [
       {
-        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
-        "action": "entered_segment",
-        "entityType": "user"
+        "serviceAction": {
+          "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+        },
+        "inputData": {
+          "entered_user_segments": ["all_segments"]
+        }
       }
     ];
     subscribe.input.data = {
       body: {
         "url": "https://hooks.zapier.com/hooks/standard/1/2/",
         "action": "entered_segment",
-        "entityType": "user"
+        "entityType": "user",
+        "inputData": {
+          "user_segments": ["user_segment_1"]
+        }
       }
     };
     subscribe.serviceRequests[0].input = {
-      "subscriptions": [
+      "triggers": [
         {
-          "url": "https://hooks.zapier.com/hooks/standard/1/1/",
-          "action": "entered_segment",
-          "entityType": "user"
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+          },
+          "inputData": {
+            "entered_user_segments": ["all_segments"]
+          }
         },
         {
-          "url": "https://hooks.zapier.com/hooks/standard/1/2/",
-          "action": "entered_segment",
-          "entityType": "user"
+          "serviceAction": {
+            "webhook": "https://hooks.zapier.com/hooks/standard/1/2/"
+          },
+          "inputData": {
+            "entered_user_segments": ["user_segment_1"]
+          }
         }
       ]
     };
@@ -75,27 +385,27 @@ describe("Zapier Subscription Tests", () => {
 
   it("Receive New Subscription And Unable To Merge With Existing Subscriptions", () => {
     const subscribe = _.cloneDeep(testDefinition);
-    subscribe.configuration.private_settings.subscriptions = [
+    subscribe.configuration.private_settings.triggers = [
       {
-        "url": "https://hooks.zapier.com/hooks/standard/1/1/",
-        "action": "entered_segment",
-        "entityType": "user"
+        "serviceAction": {
+          "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+        },
+        "inputData": {
+          "entered_user_segments": ["all_segments"]
+        }
       }
     ];
     subscribe.input.data = {
       body: {
         "url": "https://hooks.zapier.com/hooks/standard/1/1/",
         "action": "entered_segment",
-        "entityType": "user"
+        "entityType": "user",
+        "inputData": {
+          "user_segments": ["user_segment_1"]
+        }
       }
     };
-    subscribe.serviceRequests = {
-      "localContext": expect.anything(),
-      "name": "hull",
-      "op": "settingsUpdate",
-      "input": {},
-      "result": {}
-    };
+    subscribe.serviceRequests = [];
     return harness.runTest(subscribe);
   });
 });